+++ container-engine-lib/docs/changelog.md
@@ -1,6 +1,6 @@
 # TBD
 ### Changes
-* Removes Docker container prefix `/` before use it to set a Container's name
+* Removes Docker container name prefix `/` before using it to set a Container's name
 
 # 0.4.3
 ### Features